Unify useage of a sentinel
Nit: instead of lots of places defining their own _null sentinel,
put it in Util and let the code get it from there.

At some point will improve this sentinel to get better traceability
(a repr with more information than use a memory address, for example)
but for now the simple one is fine, just don't define the same
thing locally in many places, or import it from Environment as some
did - that carries danger of import loops.

Signed-off-by: Mats Wichmann <mats@linux.com>